re-generate trajectory by using ros param
